+++ assets/data/campaign/submarkets.csv
@@ -11,3 +11,10 @@ Selling a commodity will not resolve a shortage.
 Black market transactions may catch the eye of authorities, resulting in cargo scans and reduced faction standing, unless the trades are carried out with the transponder turned off.
 
 Engaging in legitimate trade as a ""cover"" also reduces the probability of attracting the attention of authorities.",starpocalypse.submarket.RegulatedBlackMarket,graphics/factions/black_market.png,999
+regulated_generic_military,,,"Trade with $theFaction military base.
+
+Weapons and other military hardware can be bought and sold here even if they're illegal on the open market. A commission and better standing with $theFaction results in higher-grade hardware being available.
+
+Will buy locally illegal goods as part of a no-questions-asked buyback program. Will also sell normally illegal items while stability is low.
+
+Selling enough of a commodity will temporarily resolve a shortage.",com.fs.starfarer.api.impl.campaign.submarkets.MilitarySubmarketPlugin,,100
