+++ assets/starpocalypse.json
@@ -11,13 +11,10 @@
 
     ## MILITARY REGULATIONS
     # Buying or selling military wares Open Market is illegal unless you are commissioned with the owner of the market.
-    # Uses same logic as Military market except for:
+    # Uses same logic as Military market with the addition of:
     # - Legal list is applied to make item or ship always legal,
-    # - Lower stability will cause some illegal items to become legal, and
     # - All civ-grade or civ-hint ships are always legal.
     "militaryRegulations": true,
-    # Legacy behaviour - remove items and ships instead of making them illegal.
-    "legacyMilitaryRegulations": false,
     # Suspicion level will grow even with transponder off. It is still useful to sell illegal wares.
     "transparentBlackMarket": true,
     # Setting this to 0 essentially disabled the feature. Setting this to 1 makes it behave as if the transponder was on
